refactor(store): atoms, actions, selectors 폴더로 상태/액션/파생 분리
- 상품, 장바구니, 쿠폰 등 도메인별 atom/action/selector 파일 분리
