+++ src/advanced/store/atoms/cartAtoms.ts
@@ -1,0 +1,5 @@
+import { atomWithStorage } from "jotai/utils";
+import { CartItem } from "../../../types";
+
+// 장바구니 상태 atom
+export const cartAtom = atomWithStorage<CartItem[]>("cart", []);
